requirements.txt
# PowerPoint生成
python-pptx>=0.6.21

# データ取得（URL入力用）
requests>=2.28.0
